refactor: add mock.rs file and reshape
